histogramTable implemented on real data
Data is gather dynamically and is used to fill up the histogram table.
added a sortObjects function in statistical functions
Added draggable functionality to the histogram table. As well as the
necessary scripts for jquery and styles needed for it to look nice. The
style is not perfect there is something wrong  with the first column on
dragging