Added Nodes table
Added another table for the nodes.
Make table creation dynamic so that we use the same functions to
populate every table.
Fixed hovering on elements on the table so that nodes highlight nodes
and links highlight links
I need to recheck data because histograms are not plotting correctly
something is wrong.