Fixed some issues with the total data tooltips
Error in the way I was calculating the total data for the tables.
Fixed the hover functionality for the total data tooltips.
Tooltips work dynamically along all the tables.
I needed to create a tooltip per table, when I was sharing tooltips it
was getting rendered below the table not on top of it for all the
tables but the first one.