Added heatmap for periodic patterns to the netsage website
needed to add a way to sum the data of the nodes from the links.
Because the links don’t have the same time samples (some have been
gathering data for more time) I needed a way to sum the different link
data of a node at the correct times. For this I created an object that
uses as index the time and as value the value at that point, if the
index doesn’t exist I create it add put the value of the link at that
tilmestep if it exists I add the value of the link to the value at that
timestamp. After I transform it to an array to feed the heat maps. (
this makes the query slow but I don’t know how to do it differently
right now)

Fixed position of text depending on the heatmap type (year,month..)

Fixed the way I was opening new tabs, now I check if I press the Ask
Netsage Button and if so I ignore the parameters in the URL and create
a new query from the selections.