Created New menu and tab support
Question title populates with the data values passed from the tab. I
don’t populate the selectors because it is going to be difficult to
follow when we have a lot of options I will see in the future what do
we do with it.

Separated the functionality to create a form from the main js file.
createForm.js
Fixed a bit the way I was doing the logic of reading parameters, now is
done at the beginning of the queryFrom function.