Changed info texts distributed across the graphs. Changed date Parsing so that it works in safari and Firefox
dataParser - Created different info texts divs so that I can position
them along the different graphs instead of everything at the top.
heatmap - Changed the way I parsed the dates, it was giving problems in
safari and Firefox. Now website works on Safari, Chrome and Firefox, to
test on Internet Explorer
mainDashboard - Changed to way I remove the info in the dashboard so
that all info divs are erased when page reloads
main.css - changes on the positioning of the info text to position them
on top of the graphs where the information is referring to.