some style changes for SC
dataParser - added window.location.name  netsage/dashboard because in
the server that is the name of the path
heatmap - Changed the names of the links for the descriptions to make
it consistent with the map, changed the tooltip and created one for the
heatmap instead of using the one of the map. Made tooltip bigger and
text in tooltip bigger too.
histogramTable - Change the sizes of the nodes on the map when hovering
on the rows in the table ( I made the nodes bigger by default)
lineChart - changed the position of the name of the line on hover for
the lines chart
mainDashboard - I made the update on the dashboard every 5 minutes
instead of every minute.
map - Changed sizes of nodes, changed positioning and sizes of tooltips
and tooltip text
main.css - Changed background of tooltips and the padding for all the
tooltips in the different graphs.