Disabled quertymeasures and queryValue from the selectors still not working
Dashboardandmaincommon - I disabled the selectors of the querymeaureas
and queryvalues for the active version of the portal and for SC. They
will be available in the future when I have them working for all the
questions.

main.js - I added the Latency option that was missing when a new
question was created in a new tab.